#394 [feat] auth 모듈 도메인 변경
