+++ module-auth/src/main/java/com/mile/client/SocialType.java
@@ -1,4 +1,4 @@
-package com.mile.external.client;
+package com.mile.client;
 
 
 import lombok.AllArgsConstructor;
