+++ module-auth/src/main/java/com/mile/client/google/api/dto/GoogleAccessTokenResponse.java
@@ -1,10 +1,11 @@
-package com.mile.external.client.kakao.api.dto;
+package com.mile.client.google.api.dto;
+
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserProfile(
-        String accountEmail
+public record GoogleAccessTokenResponse(
+        String accessToken
 ) {
 }
