+++ module-auth/src/main/java/com/mile/client/kakao/api/dto/KakaoAccount.java
@@ -1,11 +1,10 @@
-package com.mile.external.client.google.api.dto;
-
+package com.mile.client.kakao.api.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record GoogleAccessTokenResponse(
-        String accessToken
+public record KakaoAccount(
+        KakaoUserProfile profile
 ) {
 }
