+++ module-auth/src/main/java/com/mile/client/kakao/api/dto/KakaoUserProfile.java
@@ -1,10 +1,10 @@
-package com.mile.external.client.kakao.api.dto;
+package com.mile.client.kakao.api.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoAccessTokenResponse(
-        String accessToken
+public record KakaoUserProfile(
+        String accountEmail
 ) {
 }
