+++ module-auth/src/main/java/com/mile/client/kakao/api/dto/KakaoUserResponse.java
@@ -1,10 +1,11 @@
-package com.mile.external.client.kakao.api.dto;
+package com.mile.client.kakao.api.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoAccount(
-        KakaoUserProfile profile
+public record KakaoUserResponse(
+        Long id,
+        KakaoAccount kakaoAccount
 ) {
 }
