+++ module-auth/src/main/java/com/mile/config/OpenFeignConfig.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.FeignAutoConfiguration;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableFeignClients("com.mile.external.client")
+@EnableFeignClients("com.mile.client")
 @ImportAutoConfiguration(FeignAutoConfiguration.class)
 public class OpenFeignConfig {
 }
