+++ module-domain/src/main/java/com/mile/user/repository/UserRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.mile.user.repository;
 
-import com.mile.external.client.SocialType;
+import com.mile.client.SocialType;
 import com.mile.user.domain.User;
 
 import java.util.Optional;
